Week 12b
Recreating repository with GitIgnore > Unity so I can push uploads to github in the first place without thousands of files/large files, and being unable to. 

Today in class we worked on the character looking around, however we weren't able to make it look left/right, nor jump. Preston is working on fixing it for next class period, after thanksgiving break.